Tag filtering is happy now.

+++ FlipCards/src/org/ttdc/flipcards/client/ui/TagNameEditor.java
@@ -8,10 +8,12 @@ import com.google.gwt.event.dom.client.ClickEvent;
 import com.google.gwt.uibinder.client.UiBinder;
 import com.google.gwt.uibinder.client.UiField;
 import com.google.gwt.uibinder.client.UiHandler;
+import com.google.gwt.user.client.Window;
 import com.google.gwt.user.client.rpc.AsyncCallback;
 import com.google.gwt.user.client.ui.Button;
 import com.google.gwt.user.client.ui.Composite;
 import com.google.gwt.user.client.ui.Label;
+import com.google.gwt.user.client.ui.TextBox;
 import com.google.gwt.user.client.ui.Widget;
 
 public class TagNameEditor extends Composite{
@@ -24,24 +26,56 @@ public class TagNameEditor extends Composite{
 	@UiField
 	Button deleteButton;
 	@UiField
-	Label tagNameLabel;
+	Button updateButton;
+	@UiField
+	TextBox tagNameEditor;
+	
 	
 	private Tag tag;
 	
 	public TagNameEditor(Tag tag) {
 		initWidget(uiBinder.createAndBindUi(this));
 		this.tag = tag;
-		deleteButton.setText("X");
-		tagNameLabel.setText(tag.getTagName());
+		deleteButton.setText("Delete");
+		updateButton.setText("Update");
+		tagNameEditor.setText(tag.getTagName());
 	}
 
 	@UiHandler("deleteButton")
 	void onClick(ClickEvent e) {
+//		Dialog.confirm(String,DialogCallback);
+		boolean sure = Window.confirm("Warning, are you sure. There could be words associated with this tag?");
+		
+		if(!sure){
+			return;
+		}
+		
 		FlipCards.studyWordsService.deleteTagName(tag.getTagId(), new AsyncCallback<Void>() {
-			
 			@Override
 			public void onSuccess(Void result) {
 				removeFromParent();
+			}
+			
+			@Override
+			public void onFailure(Throwable caught) {
+				FlipCards.showErrorMessage(caught.getMessage());
+			}
+		});
+	}
+	
+	@UiHandler("updateButton")
+	void onUpdateClick(ClickEvent e) {
+		String value = tagNameEditor.getText().trim();
+		if(value.isEmpty()){
+			FlipCards.showErrorMessage("Tag can't be blank");
+			return;
+		}
+		
+		FlipCards.studyWordsService.updateTagName(tag.getTagId(), value, new AsyncCallback<Tag>() {
+			@Override
+			public void onSuccess(Tag result) {
+				tag.setTagName(result.getTagName());
+				tagNameEditor.setText(result.getTagName());
 			}
 			
 			@Override
